+++ changelog.md
@@ -25,7 +25,8 @@
 * Tweak - Add Sensei compatibility
 * Tweak - Add The Events Calendar compatibility
 * Tweak - Default colours scheme updated (top menu colors)
-* Tweak - Change WooCommerce Bundle product visual on single
+* Tweak - Change WooCommerce Bundle Product visual on single
+* Tweak - Change WooCommerce Variation Swatches visual on single
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned