+++ changelog.md
@@ -27,6 +27,7 @@
 * Tweak - Default colours scheme updated (top menu colors)
 * Tweak - Change WooCommerce Bundle Product visual on single
 * Tweak - Change WooCommerce Variation Swatches visual on single
+* Tweak - Change WooCommerce Ship Multiple Addresses visual on single
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned