+++ changelog.md
@@ -28,6 +28,7 @@
 * Tweak - Change WooCommerce Bundle Product visual on single
 * Tweak - Change WooCommerce Variation Swatches visual on single
 * Tweak - Change WooCommerce Ship Multiple Addresses visual on single
+* Fix - Remove WC Shipping Multiple Addresses specific script causing issues on checkout
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned