+++ changelog.md
@@ -30,6 +30,7 @@
 * Tweak - Change WooCommerce Ship Multiple Addresses visual on single
 * Fix - Remove WC Shipping Multiple Addresses specific script causing issues on checkout
 * Tweak - Change WooCommerce Products Addon visual on single
+* Tweak - Add WooCommerce Bookings compatibility
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned