+++ changelog.md
@@ -37,6 +37,7 @@
 * Fix - Make the menu background image full-width (infinite) when it is a 100% column
 * Tweak - Add WooCommerce Subscriptions compatibility
 * Tweak - Add WooCommerce Product Retailers compatibility
+* Dev - Extended alert-danger Bootstrap class to alert-error class
 
 ### 2.0.0
 * New totally awesome version! The visual was fully redesigned